Refactor weather.js
